# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Car Rental API Documentation",
    "description": "",
    "contact": {
      "email": "kevin@mail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
    "/cars": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "get all cars",
        "description": "get full list of cars",
        "operationId": "getCars",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Car not found"
          },
          "500": {
            "description": "Interval Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "cars"
        ],
        "summary": "Add a new car",
        "description": "Add a new car",
        "operationId": "addCar",
        "requestBody": {
          "$ref": "#/components/requestBodies/Car",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cars/{id}": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "get car by id",
        "description": "get detail car",
        "operationId": "getCarsId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet that needs to get specific",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data fetch succesfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Car not found"
          },
          "500": {
            "description": "Interval Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "cars"
        ],
        "summary": "Update a car",
        "description": "Updating a car",
        "operationId": "updateCar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of car that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Car"
        },
        "responses": {
          "200": {
            "description": "Data updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Car not found"
          }
        }
      },
      "delete": {
        "tags": [
          "cars"
        ],
        "summary": "Delete a car",
        "description": "Deleting a car",
        "operationId": "deleteCar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of car that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Car is deleted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Car not found"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Car": {
        "required": [
          "name",
          "price",
          "manufactur",
          "isDriver",
          null
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "Yaris"
          },
          "manufactur": {
            "type": "string",
            "example": "Toyota"
          },
          "type": {
            "type": "string",
            "example": "Hatchback"
          },
          "licenseNumber": {
            "type": "string",
            "example": "T 1235 TI"
          },
          "seat": {
            "type": "number",
            "example": 5
          },
          "baggage": {
            "type": "number",
            "example": 50
          },
          "transmission": {
            "type": "string",
            "example": "CVT"
          },
          "year": {
            "type": "string",
            "example": 2020
          },
          "description": {
            "type": "string",
            "example": "Mobil kecil enak buat nyalip"
          },
          "isDriver": {
            "type": "boolean",
            "example": false
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "img": {
            "type": "string",
            "example": "http://google.com/yaris.jpg"
          },
          "price": {
            "type": "number",
            "example": 150000
          }
        }
      },
...
    },
    "requestBodies": {
      "Car": {
        "description": "Pet object that needs to be added to the store",
        "content": {
          "application/json": {
            "example": {
              "name": "Yaris",
              "manufactur": "Toyota",
              "type": "Hatchback",
              "licenseNumber": "T 1235 TI",
              "seat": 5,
              "baggage": 50,
              "transmission": "CVT",
              "year": 2020,
              "description": "Mobil kecil enak buat nyalip",
              "isDriver": false,
              "isAvailable": true,
              "img": "http://google.com/yaris.jpg",
              "price": 150000
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}